refactor: move all routes to back/routes
