+++ src/back/routes/badges.ts
@@ -2,10 +2,9 @@ import handleAPI from 'decentraland-gatsby/dist/entities/Route/handle'
 import routes from 'decentraland-gatsby/dist/entities/Route/routes'
 import { Request } from 'express'
 
-import { validateAddress } from '../../back/utils/validations'
+import { UserBadges } from '../../entities/Badges/types'
 import { BadgesService } from '../../services/BadgesService'
-
-import { UserBadges } from './types'
+import { validateAddress } from '../utils/validations'
 
 export default routes((router) => {
   router.get('/badges/:address/', handleAPI(getBadges))
