+++ src/back/routes/committee.ts
@@ -1,7 +1,7 @@
 import handleAPI from 'decentraland-gatsby/dist/entities/Route/handle'
 import routes from 'decentraland-gatsby/dist/entities/Route/routes'
 
-import { COMMITTEE_ADDRESSES } from './isDAOCommittee'
+import { COMMITTEE_ADDRESSES } from '../../entities/Committee/isDAOCommittee'
 
 export default routes((router) => {
   return router.get(
