+++ src/server.ts
@@ -15,24 +15,24 @@ import { register } from 'prom-client'
 import swaggerUi from 'swagger-ui-express'
 import YAML from 'yaml'
 
+import badges from './back/routes/badges'
+import budget from './back/routes/budget'
+import coauthor from './back/routes/coauthor'
+import committee from './back/routes/committee'
+import common from './back/routes/common'
+import debug from './back/routes/debug'
+import proposal from './back/routes/proposal'
+import sitemap from './back/routes/sitemap'
 import snapshot from './back/routes/snapshot'
-import badges from './entities/Badges/routes'
+import social from './back/routes/social'
+import subscription from './back/routes/subscription'
+import proposalSurveyTopics from './back/routes/surveyTopics'
+import updates from './back/routes/updates'
+import users from './back/routes/user'
+import score from './back/routes/votes'
 import bids from './entities/Bid/routes'
 import { updateGovernanceBudgets } from './entities/Budget/jobs'
-import budget from './entities/Budget/routes'
-import coauthor from './entities/Coauthor/routes'
-import committee from './entities/Committee/routes'
-import common from './entities/Common/routes'
-import debug from './entities/Debug/routes'
 import { activateProposals, finishProposal, publishBids } from './entities/Proposal/jobs'
-import proposal from './entities/Proposal/routes'
-import proposalSurveyTopics from './entities/ProposalSurveyTopics/routes'
-import sitemap from './entities/Sitemap/routes'
-import social from './entities/Social/routes'
-import subscription from './entities/Subscription/routes'
-import updates from './entities/Updates/routes'
-import users from './entities/User/routes'
-import score from './entities/Votes/routes'
 import { DiscordService } from './services/DiscordService'
 import filesystem from './utils/filesystem'
